feat(test): @blobscan/blob-storage-manager: refactor to test folder @blobscan/api: sort indexer data to keep snapshot deterministic
