+++ packages/blob-storage-manager/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "test/fixtures.ts", "vitest.config.ts"]
+  "include": ["src", "test", "vitest.config.ts"]
 }
